more refactor on types
