+++ crates/fiber-bin/src/main.rs
@@ -8,7 +8,8 @@ use fnn::ckb::contracts::TypeIDResolver;
 #[cfg(debug_assertions)]
 use fnn::ckb::contracts::{get_cell_deps, Contract};
 use fnn::ckb::{contracts::try_init_contracts_context, CkbChainActor};
-use fnn::fiber::{graph::NetworkGraph, network::init_chain_hash, network::NetworkActorMessage};
+use fnn::fiber::{graph::NetworkGraph, network::NetworkActorMessage};
+use fnn::fiber_types::init_chain_hash;
 use fnn::rpc::server::start_rpc;
 use fnn::rpc::watchtower::{
     CreatePreimageParams, CreateWatchChannelParams, RemovePreimageParams, RemoveWatchChannelParams,
