+++ crates/fiber-lib/benches/payment_benchmarks.rs
@@ -2,17 +2,17 @@
 use ckb_types::packed::OutPoint;
 use ckb_types::prelude::Entity;
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, Throughput};
-use fiber_types::{FeatureVector, HopHint};
+use fiber_types::{ChannelUpdate, FeatureVector, HopHint};
 use fnn::fiber::config::{
     DEFAULT_FINAL_TLC_EXPIRY_DELTA, DEFAULT_TLC_EXPIRY_DELTA, MAX_PAYMENT_TLC_EXPIRY_LIMIT,
 };
 use fnn::fiber::gossip::GossipMessageStore;
 use fnn::fiber::graph::{GraphChannelStat, NetworkGraph};
-use fnn::fiber::network::get_chain_hash;
 use fnn::fiber::types::{
-    new_channel_update_unsigned, new_node_announcement, ChannelAnnouncement,
-    ChannelUpdateChannelFlags, ChannelUpdateMessageFlags, Pubkey,
+    new_node_announcement, ChannelAnnouncement, ChannelUpdateChannelFlags,
+    ChannelUpdateMessageFlags, Pubkey,
 };
+use fnn::fiber_types::get_chain_hash;
 use fnn::store::Store;
 use fnn::tests::create_n_nodes_network;
 use fnn::MIN_RESERVED_CKB;
@@ -277,7 +277,7 @@ fn build_graph_for_find_path_bench(
             },
         );
 
-        store.save_channel_update(new_channel_update_unsigned(
+        store.save_channel_update(ChannelUpdate::new_unsigned(
             channel_outpoint.clone(),
             now_timestamp_as_millis_u64(),
             if node_a_is_node1 {
@@ -290,7 +290,7 @@ fn build_graph_for_find_path_bench(
             0,                        // tlc_minimum_value
             0, // tlc_fee_proportional_millionths (fee rate, should be small, not capacity!)
         ));
-        store.save_channel_update(new_channel_update_unsigned(
+        store.save_channel_update(ChannelUpdate::new_unsigned(
             channel_outpoint.clone(),
             now_timestamp_as_millis_u64(),
             if node_a_is_node1 {
