+++ crates/fiber-lib/src/fiber/types.rs
@@ -1,5 +1,4 @@
 use super::channel::ProcessingChannelError;
-use super::network::get_chain_hash;
 use crate::ckb::contracts::get_udt_whitelist;
 use anyhow::anyhow;
 use ckb_jsonrpc_types::CellOutput;
@@ -10,6 +9,7 @@ use ckb_types::{
 };
 use core::fmt::{self, Formatter};
 use fiber_sphinx::SphinxError;
+use fiber_types::get_chain_hash;
 use fiber_types::molecule_table_data_len;
 use fiber_types::protocol::AnnouncedNodeName;
 pub use fiber_types::{
@@ -865,35 +865,6 @@ pub fn new_channel_announcement_unsigned(
         ckb_key: *ckb_pubkey,
         capacity,
         udt_type_script,
-    }
-}
-
-pub fn new_channel_update_unsigned(
-    channel_outpoint: OutPoint,
-    timestamp: u64,
-    message_flags: ChannelUpdateMessageFlags,
-    channel_flags: ChannelUpdateChannelFlags,
-    tlc_expiry_delta: u64,
-    tlc_minimum_value: u128,
-    tlc_fee_proportional_millionths: u128,
-) -> ChannelUpdate {
-    // To avoid having the same timestamp for both channel updates, we will use an even
-    // timestamp number for node1 and an odd timestamp number for node2.
-    let timestamp = if message_flags.contains(ChannelUpdateMessageFlags::UPDATE_OF_NODE2) {
-        timestamp | 1u64
-    } else {
-        timestamp & !1u64
-    };
-    ChannelUpdate {
-        signature: None,
-        chain_hash: get_chain_hash(),
-        channel_outpoint,
-        timestamp,
-        message_flags,
-        channel_flags,
-        tlc_expiry_delta,
-        tlc_minimum_value,
-        tlc_fee_proportional_millionths,
     }
 }
