+++ crates/fiber-lib/src/rpc/graph.rs
@@ -1,10 +1,10 @@
 use crate::fiber::channel::ChannelActorStateStore;
 use crate::fiber::gossip::GossipMessageStore;
 use crate::fiber::graph::{ChannelUpdateInfo, NetworkGraph, NetworkGraphStateStore};
-use crate::fiber::network::get_chain_hash;
 use ckb_jsonrpc_types::{DepType, JsonBytes, OutPoint as OutPointWrapper, Script, ScriptHashType};
 use ckb_types::packed::OutPoint;
 use ckb_types::H256;
+use fiber_types::get_chain_hash;
 use fiber_types::{
     Cursor, EntityHex, Hash256, Pubkey, U128Hex, U64Hex, UdtArgInfo as ConfigUdtArgInfo,
     UdtCellDep as ConfigUdtCellDep, UdtCfgInfos as ConfigUdtCfgInfos, UdtDep as ConfigUdtDep,
