+++ crates/fiber-lib/src/tests/gen_utils.rs
@@ -6,9 +6,7 @@ use ckb_types::{packed::OutPoint, prelude::Pack};
 use secp256k1::{Keypair, PublicKey, SecretKey, XOnlyPublicKey, SECP256K1};
 
 use crate::ckb::contracts::{get_cell_deps_by_contracts, get_script_by_contract, Contract};
-use crate::fiber::types::{
-    new_channel_announcement_unsigned, new_channel_update_unsigned, new_node_announcement,
-};
+use crate::fiber::types::{new_channel_announcement_unsigned, new_node_announcement};
 use crate::fiber::{
     ChannelAnnouncement, ChannelUpdate, ChannelUpdateChannelFlags, ChannelUpdateMessageFlags,
     EcdsaSignature, FeatureVector, NodeAnnouncement, Privkey, Pubkey,
@@ -162,7 +160,7 @@ impl ChannelTestContext {
         timestamp: Option<u64>,
     ) -> ChannelUpdate {
         let timestamp = timestamp.unwrap_or(now_timestamp_as_millis_u64());
-        let mut unsigned_channel_update = new_channel_update_unsigned(
+        let mut unsigned_channel_update = ChannelUpdate::new_unsigned(
             self.channel_announcement.channel_outpoint.clone(),
             timestamp,
             ChannelUpdateMessageFlags::UPDATE_OF_NODE1,
@@ -186,7 +184,7 @@ impl ChannelTestContext {
         timestamp: Option<u64>,
     ) -> ChannelUpdate {
         let timestamp = timestamp.unwrap_or(now_timestamp_as_millis_u64());
-        let mut unsigned_channel_update = new_channel_update_unsigned(
+        let mut unsigned_channel_update = ChannelUpdate::new_unsigned(
             self.channel_announcement.channel_outpoint.clone(),
             timestamp,
             ChannelUpdateMessageFlags::UPDATE_OF_NODE2,
