+++ crates/fiber-types/Cargo.toml
@@ -8,6 +8,7 @@ description = "Core domain types for the Fiber Network"
 # Serialization
 serde = { version = "1.0", features = ["derive"] }
 serde_with = { version = "3.7", features = ["macros"] }
+once_cell = "1.19.0"
 
 # Encoding
 hex = "0.4"
